Remove files we don't need in core
Moved them to the code generator project in aws-sdk-swift
Kept Location as this is used in both project
